Modified the main kint build function. To use a routerBuilder
